requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
astor==0.8.1
astunparse==1.6.3
bcrypt==5.0.0
beam-client==0.2.190
beta9==0.1.243
betterproto-beta9==2.0.1
black==25.11.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
croniter==3.0.4
cryptography==46.0.3
decorator==5.2.1
exceptiongroup==1.3.1
fastapi==0.128.0
flatbuffers==25.12.19
gast==0.7.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.69.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==5.2.0
isort==5.13.2
Jinja2==3.1.6
joblib==1.5.3
keras==2.15.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.18.0
marshmallow_dataclass==8.6.1
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
paramiko==3.5.1
pathspec==1.0.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
starlette==0.49.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.3.0
typeguard==2.13.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.39.0
validators==0.20.0
watchdog==4.0.2
wcwidth==0.2.14
websocket-client==1.9.0
websockets==13.1
Werkzeug==3.1.4
wrapt==1.14.2
zipp==3.23.0
